Changes for 3.10 (32160)
- Updates the Model/Input classes to match the new export JSON structure
- Updates references to model structure
- Updates SevenZipSharp package
- Adds a few null checks

+++ Model/Vector3.cs
@@ -7,15 +7,15 @@ namespace LootDumpProcessor.Model
     {
         [JsonProperty("x", NullValueHandling = NullValueHandling.Ignore)]
         [JsonPropertyName("x")]
-        public double? X { get; set; }
+        public float? X { get; set; }
 
         [JsonProperty("y", NullValueHandling = NullValueHandling.Ignore)]
         [JsonPropertyName("y")]
-        public double? Y { get; set; }
+        public float? Y { get; set; }
 
         [JsonProperty("z", NullValueHandling = NullValueHandling.Ignore)]
         [JsonPropertyName("z")]
-        public double? Z { get; set; }
+        public float? Z { get; set; }
 
         public object Clone()
         {
